+++ src/advanced/components/layout/index.ts
@@ -1,0 +1,5 @@
+export { default as Header } from '@/advanced/components/layout/Header';
+export { default as HelpOverlay } from '@/advanced/components/layout/HelpOverlay';
+export { default as HelpToggle } from '@/advanced/components/layout/HelpToggle';
+export { default as MainLayout } from '@/advanced/components/layout/MainLayout';
+export { default as SectionLayout } from '@/advanced/components/layout/SectionLayout';
