+++ src/advanced/components/order/index.ts
@@ -1,0 +1,1 @@
+export { default as OrderSummary } from '@/advanced/components/order/OrderSummary';
