+++ src/advanced/components/product/index.ts
@@ -1,0 +1,1 @@
+export { default as ProductSelect } from '@/advanced/components/product/ProductSelect';
