+++ src/advanced/store/index.ts
@@ -1,0 +1,3 @@
+export { default as useCartStore } from '@/advanced/store/useCartStore';
+export { default as useLayoutStore } from '@/advanced/store/useLayoutStore';
+export { default as useProductStore } from '@/advanced/store/useProductStore';
